// workflow_fe/src/lib/api/gen/types/InvoiceResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { WorkflowUserResponseDto } from "./WorkflowUserResponseDto.ts";

export type InvoiceResponseDtoStatusEnum = "ISSUED" | "CANCELLED" | "PAID";

export type InvoiceResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  billingRequestId: string;
  /**
   * @type string
   */
  invoiceNumber: string;
  /**
   * @type string
   */
  requesterId: string;
  /**
   * @type object
   */
  requester: WorkflowUserResponseDto | null;
  /**
   * @type string
   */
  departmentId: string | null;
  /**
   * @type string
   */
  customerName: string;
  /**
   * @type string
   */
  customerEmail: string | null;
  /**
   * @type string
   */
  customerAddress: string | null;
  /**
   * @type string
   */
  title: string;
  /**
   * @type string
   */
  description: string | null;
  /**
   * @type string
   */
  amount: string;
  /**
   * @type string
   */
  currency: string;
  /**
   * @type string
   */
  dueDate: string;
  /**
   * @type string
   */
  status: InvoiceResponseDtoStatusEnum;
  /**
   * @type string
   */
  issuedAt: string;
  /**
   * @type string, date-time
   */
  cancelledAt: string | null;
  /**
   * @type string, date-time
   */
  paidAt: string | null;
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};